Moving Deferred Promise to src/util
