+++ dev/test/pool.ts
@@ -18,7 +18,7 @@ import {expect, use} from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
 
 import {ClientPool} from '../src/pool';
-import {Deferred} from './util/helpers';
+import {Deferred} from '../src/util';
 
 use(chaiAsPromised);
